+++ fotoskazka/config/contacts.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'phone' => env('CONTACT_PHONE', '+7 (___) ___-__-__'),
+    'email' => env('CONTACT_EMAIL', 'info@fotoskazka.ru'),
+];